[eventwizard] Move internal APIs to own blueprint
