+++ src/backend/api/handlers/tests/add_fms_report_archive_test.py
@@ -101,7 +101,7 @@ def test_upload_qual_rankings_success(
     file_content = create_excel_file()
     file_digest = hashlib.sha256(file_content).hexdigest()
 
-    request_path = "/api/trusted/v1/event/2019nyny/fms_reports/qual_rankings"
+    request_path = "/_eventwizard/event/2019nyny/fms_reports/qual_rankings"
 
     file_storage = FileStorage(
         stream=io.BytesIO(file_content),
@@ -109,8 +109,10 @@ def test_upload_qual_rankings_success(
         content_type="application/vnd.ms-excel",
     )
 
-    with patch("backend.api.handlers.trusted.storage_write") as mock_write:
-        with patch("backend.api.handlers.trusted.storage_get_files") as mock_get_files:
+    with patch("backend.api.handlers.eventwizard_internal.storage_write") as mock_write:
+        with patch(
+            "backend.api.handlers.eventwizard_internal.storage_get_files"
+        ) as mock_get_files:
             mock_get_files.return_value = []
 
             resp = api_client.post(
@@ -145,7 +147,7 @@ def test_upload_missing_file(
         expiration=None,
     )
 
-    request_path = "/api/trusted/v1/event/2019nyny/fms_reports/qual_rankings"
+    request_path = "/_eventwizard/event/2019nyny/fms_reports/qual_rankings"
 
     resp = api_client.post(
         request_path,
@@ -178,7 +180,7 @@ def test_upload_invalid_excel_file(
     file_content = b"This is not an Excel file"
     file_digest = hashlib.sha256(file_content).hexdigest()
 
-    request_path = "/api/trusted/v1/event/2019nyny/fms_reports/qual_rankings"
+    request_path = "/_eventwizard/event/2019nyny/fms_reports/qual_rankings"
 
     file_storage = FileStorage(
         stream=io.BytesIO(file_content),
@@ -220,7 +222,7 @@ def test_upload_wrong_permission(
     file_content = create_excel_file()
     file_digest = hashlib.sha256(file_content).hexdigest()
 
-    request_path = "/api/trusted/v1/event/2019nyny/fms_reports/qual_rankings"
+    request_path = "/_eventwizard/event/2019nyny/fms_reports/qual_rankings"
 
     file_storage = FileStorage(
         stream=io.BytesIO(file_content),
@@ -263,7 +265,7 @@ def test_upload_mismatched_digest(
     wrong_digest = hashlib.sha256(b"wrong content").hexdigest()
     correct_digest = hashlib.sha256(file_content).hexdigest()
 
-    request_path = "/api/trusted/v1/event/2019nyny/fms_reports/qual_rankings"
+    request_path = "/_eventwizard/event/2019nyny/fms_reports/qual_rankings"
 
     file_storage = FileStorage(
         stream=io.BytesIO(file_content),
@@ -305,7 +307,7 @@ def test_upload_duplicate_file_not_written(
     file_content = create_excel_file()
     file_digest = hashlib.sha256(file_content).hexdigest()
 
-    request_path = "/api/trusted/v1/event/2019nyny/fms_reports/qual_rankings"
+    request_path = "/_eventwizard/event/2019nyny/fms_reports/qual_rankings"
 
     file_storage = FileStorage(
         stream=io.BytesIO(file_content),
@@ -319,8 +321,10 @@ def test_upload_duplicate_file_not_written(
         "fms_reports/2019nyny/qual_rankings/rankings.2019-06-01 00:00:00.xlsx"
     ]
 
-    with patch("backend.api.handlers.trusted.storage_write") as mock_write:
-        with patch("backend.api.handlers.trusted.storage_get_files") as mock_get_files:
+    with patch("backend.api.handlers.eventwizard_internal.storage_write") as mock_write:
+        with patch(
+            "backend.api.handlers.eventwizard_internal.storage_get_files"
+        ) as mock_get_files:
             mock_get_files.return_value = existing_files
 
             resp = api_client.post(
@@ -358,7 +362,7 @@ def test_upload_with_metadata(
     file_content = create_excel_file()
     file_digest = hashlib.sha256(file_content).hexdigest()
 
-    request_path = "/api/trusted/v1/event/2019nyny/fms_reports/qual_rankings"
+    request_path = "/_eventwizard/event/2019nyny/fms_reports/qual_rankings"
 
     file_storage = FileStorage(
         stream=io.BytesIO(file_content),
@@ -366,8 +370,10 @@ def test_upload_with_metadata(
         content_type="application/vnd.ms-excel",
     )
 
-    with patch("backend.api.handlers.trusted.storage_write") as mock_write:
-        with patch("backend.api.handlers.trusted.storage_get_files") as mock_get_files:
+    with patch("backend.api.handlers.eventwizard_internal.storage_write") as mock_write:
+        with patch(
+            "backend.api.handlers.eventwizard_internal.storage_get_files"
+        ) as mock_get_files:
             mock_get_files.return_value = []
 
             resp = api_client.post(
@@ -401,7 +407,7 @@ def test_upload_not_authenticated(ndb_stub, api_client: Client) -> None:
     file_content = create_excel_file()
     file_digest = hashlib.sha256(file_content).hexdigest()
 
-    request_path = "/api/trusted/v1/event/2019nyny/fms_reports/qual_rankings"
+    request_path = "/_eventwizard/event/2019nyny/fms_reports/qual_rankings"
 
     file_storage = FileStorage(
         stream=io.BytesIO(file_content),
